Add files via upload
Open the Spring Initializr (https://start.spring.io/).
Select the Spring Boot version, project name(com.projectName) ,language, java version and package.
Add dependencies like Spring web, Spring data JPA, Spring boot dev tools and Mysql next Generate files.
Extract the jar file.
Import the project.
File -> Import -> Existing Maven Projects -> Next -> Browse -> Select the folder spring- spring-boot-application-run -> Select Folder -> Finish.
It takes time to import the project. When the project imports successfully, we can see it in the Package Explorer section of the IDE.
We see that two files are automatically created, one is pom.xml, and the other is src/main/java and src/main/resources.
Inside the src/main/java automatically one package with class is created.
we should pass the packages inside the in-built pacakage class.
we should create a database for the project is Insurance Management Platform.
Inside the src/main/resources and the application.properties present, we add the Supportingfiles from the github should add it, and we need to add the Mysql database name, url and password inside the properties and save it.
Inside the pom.xml file contains the all the dependencies, application name, Spring Boot version, group name, artifact, and other plugins.
The main class is a class that contains the main() method. It starts the Spring ApplicationContext. 
It is the class that we run for the execution of the application.
Add the package are : 
           DTO (data transfer object)
           Repository
           DAO (data access object)
           Controller
Dto :-  data transfer object should be used in the dto package and the classes should create inside the dto package.
Repository :-  JPA Repository are in the form of Interface class and access the In-built package of the repository class.
Dao :-  The Objects should be accessed from the dto package and in-built classes are accessed in the JPA repository classes and Create the objects for the methods and the methods such as insertAll(), deleteAll(), save(), getById() and update() etc.....
Controller :-  Pre-defined methods such as GetMapping, PostMapping, PutMapping and DeleteMapping etc... methods are used in the Controller class to access the Object class objects.
After writing the code must and should save the all class (ctrl+s).
Then run the code on the Main-method of automatically created class by the SpringBootProject.
  ==> Now , we can go for Postman 
  go to workplace and move to my workspaces.
 Select the methods like POST,GET,PUT,DELETE and UPDATE etc... to select the method.
 Then select the Complete Url path from Postman to the Controller class.
 Select body --> raw --> text( TEXT, JAVASCRIPT, XML, JSON, HTML ).
We should select the JSON then passed the values inside double codes and the send the request.
Before sending the send request, we should RUN the code once, then send the request inside postman.
Now, given values are present in the MYSQL DATABASE .
Simillary, we do remaining operations in the same format.
These is about Spring Boot Project on Insurance Management Platform..